Simplifying and cleaning up includes
