+++ include/glaze/api/std/string.hpp
@@ -4,6 +4,7 @@
 #pragma once
 
 #include <string>
+#include <string_view>
 
 #include "glaze/api/name.hpp"
