--- include/glaze/util/string_view.hpp
@@ -1,11 +1,0 @@
-// Glaze Library
-// For the license information refer to glaze.hpp
-
-#pragma once
-
-#include <string_view>
-
-namespace glz
-{
-   using sv = std::string_view;
-}
