+++ tests/json_test/json_test.cpp
@@ -24,13 +24,8 @@
 #include "glaze/api/impl.hpp"
 #include "glaze/file/hostname_include.hpp"
 #include "glaze/file/raw_or_file.hpp"
-#include "glaze/json/json_ptr.hpp"
-#include "glaze/json/prettify.hpp"
-#include "glaze/json/ptr.hpp"
-#include "glaze/json/quoted.hpp"
-#include "glaze/json/read.hpp"
+#include "glaze/json.hpp"
 #include "glaze/json/study.hpp"
-#include "glaze/json/write.hpp"
 #include "glaze/record/recorder.hpp"
 #include "glaze/util/poly.hpp"
 #include "glaze/util/progress_bar.hpp"
